create bird component from bird_list to split responsibilities
